# log: fix remaining misuse of runtime.Callers/runtime.FuncForPC