@RaduBerinde 2019-02-28
Does this also fix #34461?